requirements.txt
numpy>=1.23
Pillow>=9.5
opencv-python-headless>=4.7
